skiplevel/pyproject.toml
# TODO: Configure project dependencies and build settings

[project]
name = "skiplevel"
version = "0.1.0"
description = "AI-powered skip-level meeting assistant"
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.7,<0.2",
    "langchain-openai>=0.0.5",
    "langgraph>=0.0.15",
    "chainlit>=0.7.700",
    "openai>=1.0.0",
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "pydantic>=1.10.8,<2.0.0",
    "tqdm>=4.66.0",
    "rich>=13.0.0",
    "urllib3<2.0.0",
    "tiktoken>=0.6.0",  # Newer version with pre-built wheels
    "grandalf>=0.7"  # For ASCII graph visualization
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["skiplevel"]
